Closed a cross-site request-forgery vulnerability in GO.
Previously, links to the functions.php script could be directly loaded from
a new session. This meant that if a user already logged into CAS was tricked
into loading the URL (such by putting the URL as the source of an image)
their account would perform the specified function without them knowing.