Feature to disable flag button if user has already submitted
Added a feature that Adam suggested to disable the flag button
on the info page if the currently logged in user has already
flagged that code. Added an array to the session var in go.php
to hold this info and adjusted flag.php and info.php approporaitely.
Still have some messy var_dumps in flag.php and still want to
account for anon users to prevent spamming of button.