Added the admin screen for flagged as inappropriate codes
- Set up a script to fetch info from the database related
to flagged go codes and display them in a table with some
features such as the ability to visit the info page by
clicking the code, or to visit the URL. Also added buttons
for "delete" and "details" for the flags related to each
code. Still need to add functionality to these buttons.
- Made a few changes to the style sheet for the flag admin
table.