Added functionality to the clear/info buttons on flag admin
- Added a handler for the info/details button that appears
for each code in the flag admin screen. The brings up a window
(currently in the same window) with additional info on the
flags for that particular code.
- Added a handler for the clear/delete button that appears
for each code in the flag admin screen. It checks for xss
and removes the rows from the DB table corresponding to that
code.
- Added these items to the go.php session and admin pages
arrays where appropriate.
- Changed the style sheet to accommodate changes to the button
names on the flag admin. Changed them from "delete" and "details"
to "clear flags" and "info" so they are distinguished from each
other more in length and pronunciation (so as to help prevent
accidental deletion when trying to access details)
- Added footer to pages that include header.

Still need to get the info button to open in a new window properly.
It was working and then stopped.