Switch in config to disable flag button for anon users and misc tweaks
- Added code to hide the flag as inappropriate button for anon
users if ANON_FLAGGING is set to false in the config.php
- Added a tab to the admin menu for "flags" that checks to see
if the current user is logged in and a go admin before display.
Adjusted the admin files to include functions.php and the
admin_nav as appropriate.
- Moved the logging of which codes have been flagged in the
current session outside of the "is user authenticated" check
so it applies to anon users session as well.
- Some style changes to the details page.