Changes to the editing interface, in progress
- Changed the way the my_codes.php works for superadmins. Now SAs
see the codes they are individual admins of and have the option of
choosing a link that lets them see all codes.
- new function "isAdmin" checks to see if the currently logged in
user is an admin of a code. Takes (code, institution).
- Added new pages to the list of pages that require authenticaiton
in go.php.
- jQuery in update2.js now provides functionality on the update2.php (temp name)
to allow adding and deleting of aliases and admins as well as some
minor validation.
- Some small changes to the update form in update2.php
- update_process.php now does some processing on the submitted
update form. It will update url, description, and public in the
db on submit.