Flag admin not dependent on code table
- Previously to display the flags, the flag admin script queried the
code table to get the url information. This meant that if the code was
deleted after a flag was set, it would break the flag display and we
wouldn't be able to get info about the flag for the deleted code. I
added a URL field to the flag table and stort the url during the setting
of a flag so the code table no longer needs to be queried.
- Also made some small wording changes to more consistenly refer to
codes as "shortcuts" to the user.