Input validation, failed field highlighting and form data preservation
- Probably should have split up these commits. :P
- Adjusted the way that validation happens. Added a static method of
code that allows for the validation of code, url, and description
- Converted the current setName methods of code and alias to use the
"new" validation.
- Added code to repopulate fields with unsubmitted data on failed
validation so the user does not lose their in-progress edits.
- Apply a class to fields that fail validation so the user can see
what field is in error.
- Success and failure of actions generates messages that are displayed
at the page the user is redirected to.
- Pass form_url to the processing script so it can send
the user back to the form they came from rather than the originating
page on failed validation.
- Added the ability to revert changes back to defaults when editing
since edited values are preserved in case you want to start from
scratch.